If I use only one socket for both the camera and the drive, the drive often goes to la-la-land.
In fact, it SYSTEMATICALLY goes to la-la-land! Whether the
camera is activated or not (and not at the same time)!

I suspect that the longer the if...else then the longer
the execution time and the more likely crash occurs

So, this code actually expects all camera msg to be on their own
socket, 5002. Let us see if it works after the html is retrofitted.

Right now the drive WORKS.